requirements.txt
amazon-textract-response-parser==1.0.3
blinker==1.9.0
boto3==1.37.26
botocore==1.37.26
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
exceptiongroup==1.2.2
Flask==3.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
marshmallow==3.26.1
openpyxl==3.1.5
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
tomli==2.2.1
urllib3==2.3.0
Werkzeug==3.1.3
python-dotenv==1.1.0
mysql-connector-python==9.2.0